Branch elvin

+++ controllers/apart.js
@@ -1,38 +1,25 @@
 const Apart = require("../models/apart");
-module.exports = {
-    save,
-    getAparts,
-    deleteOneApart,
-};
+
 const { response } = require("express");
-const { ObjectId } = require("mongodb");
 
 const save = async (req, res = response) => {
-  const { square_meter, apart_number } = req.body;
-  try {
-    let apart = new Apart();
-    apart.square_meter = square_meter;
-    apart.apart_number = apart_number;
-    await apart.save();
-
-    await Apart.findById(ObjectId(apart._id))
-      .then((get_apart) => {
-        return res.status(200).json({
-          status: "success",
-          msg: "Agregado correctamente",
-          apart: get_apart,
-        });
-      })
-      .catch((err) => {
-        console.log(err);
-        return res.status(500).send({
-          msg: "Por favor hable con el administrador",
-        });
+  if (req.user.role === "ROLE_ADMINISTRATOR") {
+    const { square_meter, apart_number } = req.body;
+    try {
+      let apart = new Apart();
+      apart.square_meter = square_meter;
+      apart.apart_number = apart_number;
+      apart.save();
+    } catch (error) {
+      return res.status(500).json({
+        status: "error",
+        msg: "Por favor hable con el administrador",
       });
-  } catch (error) {
-    return res.status(500).json({
+    }
+  } else {
+    return res.status(400).send({
       status: "error",
-      msg: "Por favor hable con el administrador",
+      msg: "No tienes permisos en la plataforma",
     });
   }
 };
@@ -53,26 +40,33 @@ const getAparts = (req, res = response) => {
   });
 };
 const deleteOneApart = (req, res = response) => {
-  let apartId = req.body.id;
-  Apart.findOneAndDelete({ _id: apartId }, (err, apart) => {
-    if (err) {
-      return res.status(500).send({
-        status: "error",
-        msg: "Error al solicitar la peticion",
+  if (req.user.role === "ROLE_ADMINISTRATOR") {
+    let apartId = req.body.id;
+    Apart.findOneAndDelete({ _id: apartId }, (err, apart) => {
+      if (err) {
+        return res.status(500).send({
+          status: "error",
+          msg: "Error al solicitar la peticion",
+        });
+      }
+      if (!apart) {
+        return res.status(404).send({
+          status: "error",
+          msg: "No se ha borrado el aparto",
+        });
+      }
+      return res.status(200).json({
+        status: "success",
+        msg: "Borrado de forma exitosa",
+        apart: apart,
       });
-    }
-    if (!apart) {
-      return res.status(404).send({
-        status: "error",
-        msg: "No se ha borrado el aparto",
-      });
-    }
-    return res.status(200).json({
-      status: "success",
-      msg: "Borrado de forma exitosa",
-      apart: apart,
     });
-  });
+  } else {
+    return res.status(400).send({
+      status: "error",
+      msg: "No tienes permisos en la plataforma",
+    });
+  }
 };
 
 module.exports = {
